fix: upgrade remaining audio paths from 16-bit to 24-bit
- IntroRecording.jsx: Float32Array chunks instead of Int16Array
- MicrophoneTest.jsx: Float32Array chunks instead of Int16Array
  (both now consistent with the updated pcm-worklet.js)
- userController.js: add -sample_fmt s32 to intro recording and
  language application ffmpeg FLAC conversions, matching the fix
  already applied to phraseController.js

+++ voicechat-frontend/src/components/call/SystemCheck/MicrophoneTest.jsx
@@ -89,7 +89,7 @@ export default function MicrophoneTest({ onSuccess }) {
             
             audioChunksRef.current = [];
             workletNode.port.onmessage = (e) => {
-                audioChunksRef.current.push(new Int16Array(e.data));
+                audioChunksRef.current.push(new Float32Array(e.data));
             };
             
             const gain = audioCtx.createGain();
@@ -137,7 +137,7 @@ export default function MicrophoneTest({ onSuccess }) {
 
         let totalLength = 0;
         for (const arr of audioChunksRef.current) totalLength += arr.length;
-        const combined = new Int16Array(totalLength);
+        const combined = new Float32Array(totalLength);
         let offset = 0;
         for (const arr of audioChunksRef.current) {
             combined.set(arr, offset);
